+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Column to public domains list which shows the owner of the domain.
+
 ## [1.2.0] - 2021-07-06
 
 ### Added